feat(map): persist map layer visibility state
This commit introduces the ability to persist the visibility state of map layers.
When a layer's visibility is toggled, its URI is added to or removed from a set of hidden layer URIs in SharedPreferences.
When layers are loaded, their visibility is set based on whether their URI is present in this set.
Additionally, when a layer is removed, its URI is also removed from the set of hidden layers.
The `MapViewModel` now clears loaded layer data in `onDispose` to ensure a clean state.
A new `StringSetPrefDelegate` is added to handle SharedPreferences for Sets of Strings.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>